feat: Gestion des exports JSON par type de diplôme
